+++ engine/src/execution_kernels/kernel.cpp
@@ -77,11 +77,6 @@ ral::execution::task_result kernel::process(std::vector<std::unique_ptr<ral::fra
 		cudaStream_t stream,
     const std::map<std::string, std::string>& args){
 
-    // TODO: figure out if this can be re enabled;
-    // if (this->has_limit_ && output->get_num_rows_added() >= this->limit_rows_) {
-    //     return;
-    // }
-
     if(inputs.size()==0){
         return {ral::execution::task_status::SUCCESS, std::string(), std::vector< std::unique_ptr<ral::frame::BlazingTable> > ()};
     }